requirements.txt
alembic==0.9.6
awscli==1.14.5
blinker==1.4
boto3==1.4.8
botocore==1.8.9
certifi==2017.11.5
chardet==3.0.4
click==6.7
codeclimate-test-reporter==0.2.3
colorama==0.3.7
coverage==4.4.2
cssmin==0.2.0
docutils==0.14
enum-compat==0.0.2
eventlet==0.21.0
Flask==0.12.2
Flask-Assets==0.12
Flask-Caching==1.3.3
Flask-Login==0.4.1
Flask-Markdown==0.3
Flask-Migrate==2.1.1
Flask-Moment==0.5.2
Flask-PageDown==0.2.2
Flask-Script==2.0.6
Flask-SocketIO==2.9.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
geojson==2.3.0
greenlet==0.4.12
hybrid==0.1.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsmin==2.2.2
Mako==1.0.7
Markdown==2.6.9
MarkupSafe==1.0
mysqlclient==1.3.12
nltk==3.2.5
numpy==1.13.3
onetimepass==1.0.1
pandas==0.21.0
pyasn1==0.4.2
pycountry==17.9.23
PyMySQL==0.7.11
python-dateutil==2.6.1
python-editor==1.0.3
python-engineio==2.0.1
python-http-client==3.0.0
python-socketio==1.8.3
pytz==2017.3
PyYAML==3.12
raven==6.3.0
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.12
scikit-learn==0.19.1
scipy==1.0.0
sendgrid==5.3.0
six==1.11.0
sklearn==0.0
SQLAlchemy==1.1.15
textblob==0.15.0
urllib3==1.22
webassets==0.12.1
Werkzeug==0.12.2
WTForms==2.1
